+++ GRDB/Core/Database.swift
@@ -1037,7 +1037,7 @@ struct ColumnInfo : RowConvertible {
         name = row.value(named: "name")
         type = row.value(named: "type")
         notNull = row.value(named: "notnull")
-        defaultDatabaseValue = row.databaseValue(named: "dflt_value")!
+        defaultDatabaseValue = row.value(named: "dflt_value")
         primaryKeyIndex = row.value(named: "pk")
     }
 }