+++ Tests/Public/Core/Row/DetachedRowTests.swift
@@ -130,11 +130,11 @@ class DetachedRowTests: GRDBTestCase {
             dbQueue.inDatabase { db in
                 let row = Row.fetchOne(db, "SELECT NULL, 1, 1.1, 'foo', x'53514C697465'")!
                 
-                guard case .null = row.databaseValue(atIndex: 0).storage else { XCTFail(); return }
-                guard case .int64(let int64) = row.databaseValue(atIndex: 1).storage, int64 == 1 else { XCTFail(); return }
-                guard case .double(let double) = row.databaseValue(atIndex: 2).storage, double == 1.1 else { XCTFail(); return }
-                guard case .string(let string) = row.databaseValue(atIndex: 3).storage, string == "foo" else { XCTFail(); return }
-                guard case .blob(let data) = row.databaseValue(atIndex: 4).storage, data == "SQLite".data(using: .utf8) else { XCTFail(); return }
+                guard case .null = (row.value(atIndex: 0) as DatabaseValue).storage else { XCTFail(); return }
+                guard case .int64(let int64) = (row.value(atIndex: 1) as DatabaseValue).storage, int64 == 1 else { XCTFail(); return }
+                guard case .double(let double) = (row.value(atIndex: 2) as DatabaseValue).storage, double == 1.1 else { XCTFail(); return }
+                guard case .string(let string) = (row.value(atIndex: 3) as DatabaseValue).storage, string == "foo" else { XCTFail(); return }
+                guard case .blob(let data) = (row.value(atIndex: 4) as DatabaseValue).storage, data == "SQLite".data(using: .utf8) else { XCTFail(); return }
             }
         }
     }
@@ -145,11 +145,11 @@ class DetachedRowTests: GRDBTestCase {
             dbQueue.inDatabase { db in
                 let row = Row.fetchOne(db, "SELECT NULL AS \"null\", 1 AS \"int64\", 1.1 AS \"double\", 'foo' AS \"string\", x'53514C697465' AS \"blob\"")!
                 
-                guard case .null = row.databaseValue(named: "null")!.storage else { XCTFail(); return }
-                guard case .int64(let int64) = row.databaseValue(named: "int64")!.storage, int64 == 1 else { XCTFail(); return }
-                guard case .double(let double) = row.databaseValue(named: "double")!.storage, double == 1.1 else { XCTFail(); return }
-                guard case .string(let string) = row.databaseValue(named: "string")!.storage, string == "foo" else { XCTFail(); return }
-                guard case .blob(let data) = row.databaseValue(named: "blob")!.storage, data == "SQLite".data(using: .utf8) else { XCTFail(); return }
+                guard case .null = (row.value(named: "null") as DatabaseValue).storage else { XCTFail(); return }
+                guard case .int64(let int64) = (row.value(named: "int64") as DatabaseValue).storage, int64 == 1 else { XCTFail(); return }
+                guard case .double(let double) = (row.value(named: "double") as DatabaseValue).storage, double == 1.1 else { XCTFail(); return }
+                guard case .string(let string) = (row.value(named: "string") as DatabaseValue).storage, string == "foo" else { XCTFail(); return }
+                guard case .blob(let data) = (row.value(named: "blob") as DatabaseValue).storage, data == "SQLite".data(using: .utf8) else { XCTFail(); return }
             }
         }
     }
@@ -198,9 +198,9 @@ class DetachedRowTests: GRDBTestCase {
             let dbQueue = try makeDatabaseQueue()
             dbQueue.inDatabase { db in
                 let row = Row.fetchOne(db, "SELECT 'foo' AS nAmE")!
-                XCTAssertEqual(row.databaseValue(named: "name"), "foo".databaseValue)
-                XCTAssertEqual(row.databaseValue(named: "NAME"), "foo".databaseValue)
-                XCTAssertEqual(row.databaseValue(named: "NaMe"), "foo".databaseValue)
+                XCTAssertEqual(row.value(named: "name") as DatabaseValue, "foo".databaseValue)
+                XCTAssertEqual(row.value(named: "NAME") as DatabaseValue, "foo".databaseValue)
+                XCTAssertEqual(row.value(named: "NaMe") as DatabaseValue, "foo".databaseValue)
                 XCTAssertEqual(row.value(named: "name") as String, "foo")
                 XCTAssertEqual(row.value(named: "NAME") as String, "foo")
                 XCTAssertEqual(row.value(named: "NaMe") as String, "foo")
@@ -213,9 +213,9 @@ class DetachedRowTests: GRDBTestCase {
             let dbQueue = try makeDatabaseQueue()
             dbQueue.inDatabase { db in
                 let row = Row.fetchOne(db, "SELECT 1 AS name, 2 AS NAME")!
-                XCTAssertEqual(row.databaseValue(named: "name"), 1.databaseValue)
-                XCTAssertEqual(row.databaseValue(named: "NAME"), 1.databaseValue)
-                XCTAssertEqual(row.databaseValue(named: "NaMe"), 1.databaseValue)
+                XCTAssertEqual(row.value(named: "name") as DatabaseValue, 1.databaseValue)
+                XCTAssertEqual(row.value(named: "NAME") as DatabaseValue, 1.databaseValue)
+                XCTAssertEqual(row.value(named: "NaMe") as DatabaseValue, 1.databaseValue)
                 XCTAssertEqual(row.value(named: "name") as Int, 1)
                 XCTAssertEqual(row.value(named: "NAME") as Int, 1)
                 XCTAssertEqual(row.value(named: "NaMe") as Int, 1)
@@ -230,7 +230,7 @@ class DetachedRowTests: GRDBTestCase {
                 let row = Row.fetchOne(db, "SELECT 'foo' AS name")!
                 
                 XCTAssertFalse(row.hasColumn("missing"))
-                XCTAssertTrue(row.databaseValue(named: "missing") == nil)
+                XCTAssertTrue(row.value(named: "missing") as DatabaseValue? == nil)
                 XCTAssertTrue(row.value(named: "missing") == nil)
             }
         }