+++ CHANGELOG.md
@@ -3,13 +3,23 @@ Release Notes
 
 ## Next Release
 
-**Fixed**
-
-- Restore support for SQLite [pre-update hooks](https://github.com/groue/GRDB.swift#support-for-sqlite-pre-update-hooks)
-
 **New**
 
 - Upgrade custom SQLite builds to [v3.14.1](http://www.sqlite.org/changes.html).
+
+**Fixed**
+
+- Restore support for SQLite [pre-update hooks](https://github.com/groue/GRDB.swift#support-for-sqlite-pre-update-hooks)
+- `DatabaseValue.fromDatabaseValue()` returns `.Null` for NULL input, instead of nil (fixes [#119](https://github.com/groue/GRDB.swift/issues/119))
+
+**Breaking Change**
+
+- `Row.databaseValue(atIndex:)` and `Row.databaseValue(named:)` have been removed. Use `value(atIndex:)` and `value(named:)` instead:
+    
+    ```diff
+    -let dbv = row.databaseValue(atIndex: 0)
+    +let dbv: DatabaseValue = row.value(atIndex: 0)
+    ```
 
 
 ## 0.82.1
@@ -338,7 +348,30 @@ Released September 11, 2016
      }
     ```
 
-## 0.81.0
+
+## 0.81.1 (Swift 2.3)
+
+Released September 16, 2016
+
+**New**
+
+- Upgrade custom SQLite builds to v3.14.1.
+
+**Fixed**
+
+- `DatabaseValue.fromDatabaseValue()` returns `.Null` for NULL input, instead of nil (fixes [#119](https://github.com/groue/GRDB.swift/issues/119))
+
+**Breaking Change**
+
+- `Row.databaseValue(atIndex:)` and `Row.databaseValue(named:)` have been removed. Use `value(atIndex:)` and `value(named:)` instead:
+    
+    ```diff
+    -let dbv = row.databaseValue(atIndex: 0)
+    +let dbv: DatabaseValue = row.value(atIndex: 0)
+    ```
+    
+
+## 0.81.0 (Swift 2.3)
 
 Released September 10, 2016
 
@@ -347,7 +380,7 @@ Released September 10, 2016
 - Swift 2.3
 
 
-## 0.80.2
+## 0.80.2 (Swift 2.2)
 
 Released September 9, 2016